+++ changelog.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on http://keepachangelog.com/en/1.0.0/
 and this project adheres to http://semver.org/spec/v2.0.0.html.
 
+## [unreleased]
+
+- Add flag -ss, --sort-skills by-experience or by-name
+
 ## [0.1.0] - 2022-02-05
 
 - Limit number of visible projects and technical skills